+++ src/main/java/com/example/ApiFilmes/repository/FilmeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilmeRepository extends JpaRepository<Filme,Long> {
+public interface FilmeRepository extends JpaRepository<Filme,Integer> {
 }
